refactor(85381): Reorganiza pastas da Gestão de Usuários
Na gestão de usuários:
- Move os componentes para uma pasta "components"
- Move o GestaoDeUsuariosService para uma pasta services
- Move o arquivo CSS para uma pasta style
