+++ src/componentes/Globais/GestaoDeUsuarios/components/GestaoDeUsuariosMain.js
@@ -1,8 +1,8 @@
 import React from "react";
-import {useGruposAcesso} from "./hooks/useGruposAcesso";
+import {useGruposAcesso} from "../hooks/useGruposAcesso";
 import {GruposAcessoInfo} from "./GruposAcessoInfo";
 import {ListaUsuarios} from "./ListaUsuarios";
-import {useUsuarios} from "./hooks/useUsuarios";
+import {useUsuarios} from "../hooks/useUsuarios";
 import {BarraTopoLista} from "./BarraTopoLista";
 import {FormFiltros} from "./FormFiltros";
 import {Paginacao} from "./Paginacao";
